revam elem and text nodes
