+++ internal/html/decoder.go
@@ -1,177 +1,168 @@
 package html
 
-import (
-	"bytes"
-	"strings"
+// func decodeNodes(s string) (node, error) {
+// 	d := &decoder{
+// 		tokenizer: html.NewTokenizer(bytes.NewBufferString(s)),
+// 	}
+// 	return d.decode()
+// }
 
-	"github.com/google/uuid"
-	"golang.org/x/net/html"
-	"golang.org/x/net/html/atom"
-)
+// type decoder struct {
+// 	tokenizer   *html.Tokenizer
+// 	decodingSVG bool
+// }
 
-func decodeNodes(s string) (node, error) {
-	d := &decoder{
-		tokenizer: html.NewTokenizer(bytes.NewBufferString(s)),
-	}
-	return d.decode()
-}
+// func (d *decoder) decode() (node, error) {
+// 	switch d.tokenizer.Next() {
+// 	case html.TextToken:
+// 		return d.decodeText()
 
-type decoder struct {
-	tokenizer   *html.Tokenizer
-	decodingSVG bool
-}
+// 	case html.SelfClosingTagToken:
+// 		return d.decodeSelfClosingElem()
 
-func (d *decoder) decode() (node, error) {
-	switch d.tokenizer.Next() {
-	case html.TextToken:
-		return d.decodeText()
+// 	case html.StartTagToken:
+// 		return d.decodeElem()
 
-	case html.SelfClosingTagToken:
-		return d.decodeSelfClosingElem()
+// 	case html.EndTagToken:
+// 		return d.closeElem()
 
-	case html.StartTagToken:
-		return d.decodeElem()
+// 	case html.ErrorToken:
+// 		return nil, d.tokenizer.Err()
 
-	case html.EndTagToken:
-		return d.closeElem()
+// 	default:
+// 		// Nothing we care about, decode the next node.
+// 		return d.decode()
+// 	}
+// }
 
-	case html.ErrorToken:
-		return nil, d.tokenizer.Err()
+// func (d *decoder) decodeText() (node, error) {
+// 	text := string(d.tokenizer.Text())
+// 	text = strings.TrimSpace(text)
 
-	default:
-		// Nothing we care about, decode the next node.
-		return d.decode()
-	}
-}
+// 	if len(text) == 0 {
+// 		// Text is empty, decode the next node.
+// 		return d.decode()
+// 	}
 
-func (d *decoder) decodeText() (node, error) {
-	text := string(d.tokenizer.Text())
-	text = strings.TrimSpace(text)
+// 	return &textNode{
+// 		id:   uuid.New().String(),
+// 		text: text,
+// 	}, nil
+// }
 
-	if len(text) == 0 {
-		// Text is empty, decode the next node.
-		return d.decode()
-	}
+// func (d *decoder) decodeSelfClosingElem() (node, error) {
+// 	name, hasAttr := d.tokenizer.TagName()
+// 	tagName := string(name)
 
-	return &textNode{
-		id:   uuid.New().String(),
-		text: text,
-	}, nil
-}
+// 	if isCompoTagName(tagName, d.decodingSVG) {
+// 		return d.decodeCompo(tagName, hasAttr)
+// 	}
 
-func (d *decoder) decodeSelfClosingElem() (node, error) {
-	name, hasAttr := d.tokenizer.TagName()
-	tagName := string(name)
+// 	return &elemNode{
+// 		id:      uuid.New().String(),
+// 		tagName: tagName,
+// 		attrs:   d.decodeAttrs(hasAttr),
+// 	}, nil
+// }
 
-	if isCompoTagName(tagName, d.decodingSVG) {
-		return d.decodeCompo(tagName, hasAttr)
-	}
+// func (d *decoder) decodeAttrs(hasAttr bool) map[string]string {
+// 	if !hasAttr {
+// 		return nil
+// 	}
 
-	return &elemNode{
-		id:      uuid.New().String(),
-		tagName: tagName,
-		attrs:   d.decodeAttrs(hasAttr),
-	}, nil
-}
+// 	attrs := make(map[string]string)
+// 	for {
+// 		name, val, moreAttr := d.tokenizer.TagAttr()
+// 		attrs[string(name)] = string(val)
 
-func (d *decoder) decodeAttrs(hasAttr bool) map[string]string {
-	if !hasAttr {
-		return nil
-	}
+// 		if !moreAttr {
+// 			break
+// 		}
+// 	}
+// 	return attrs
+// }
 
-	attrs := make(map[string]string)
-	for {
-		name, val, moreAttr := d.tokenizer.TagAttr()
-		attrs[string(name)] = string(val)
+// func (d *decoder) decodeElem() (node, error) {
+// 	name, hasAttr := d.tokenizer.TagName()
+// 	tagName := string(name)
 
-		if !moreAttr {
-			break
-		}
-	}
-	return attrs
-}
+// 	if isCompoTagName(tagName, d.decodingSVG) {
+// 		return d.decodeCompo(tagName, hasAttr)
+// 	}
 
-func (d *decoder) decodeElem() (node, error) {
-	name, hasAttr := d.tokenizer.TagName()
-	tagName := string(name)
+// 	if tagName == "svg" {
+// 		d.decodingSVG = true
+// 	}
 
-	if isCompoTagName(tagName, d.decodingSVG) {
-		return d.decodeCompo(tagName, hasAttr)
-	}
+// 	elem := &elemNode{
+// 		id:      uuid.New().String(),
+// 		tagName: tagName,
+// 		attrs:   d.decodeAttrs(hasAttr),
+// 	}
 
-	if tagName == "svg" {
-		d.decodingSVG = true
-	}
+// 	if isVoidElem(tagName) {
+// 		return elem, nil
+// 	}
 
-	elem := &elemNode{
-		id:      uuid.New().String(),
-		tagName: tagName,
-		attrs:   d.decodeAttrs(hasAttr),
-	}
+// 	for {
+// 		child, err := d.decode()
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		if child == nil {
+// 			break
+// 		}
+// 		elem.appendChild(child)
+// 	}
+// 	return elem, nil
+// }
 
-	if isVoidElem(tagName) {
-		return elem, nil
-	}
+// func (d *decoder) closeElem() (node, error) {
+// 	tagName, _ := d.tokenizer.TagName()
+// 	if string(tagName) == "svg" {
+// 		d.decodingSVG = false
+// 	}
+// 	return nil, nil
+// }
 
-	for {
-		child, err := d.decode()
-		if err != nil {
-			return nil, err
-		}
-		if child == nil {
-			break
-		}
-		elem.appendChild(child)
-	}
-	return elem, nil
-}
+// func (d *decoder) decodeCompo(tagName string, hasAttr bool) (node, error) {
+// 	return &compoNode{
+// 		id:     uuid.New().String(),
+// 		name:   tagName,
+// 		fields: d.decodeAttrs(hasAttr),
+// 	}, nil
+// }
 
-func (d *decoder) closeElem() (node, error) {
-	tagName, _ := d.tokenizer.TagName()
-	if string(tagName) == "svg" {
-		d.decodingSVG = false
-	}
-	return nil, nil
-}
+// func isHTMLTagName(tagName string) bool {
+// 	return atom.Lookup([]byte(tagName)) != 0
+// }
 
-func (d *decoder) decodeCompo(tagName string, hasAttr bool) (node, error) {
-	return &compoNode{
-		id:     uuid.New().String(),
-		name:   tagName,
-		fields: d.decodeAttrs(hasAttr),
-	}, nil
-}
+// func isCompoTagName(tagName string, decodingSVG bool) bool {
+// 	if decodingSVG {
+// 		return false
+// 	}
+// 	return !isHTMLTagName(tagName)
+// }
 
-func isHTMLTagName(tagName string) bool {
-	return atom.Lookup([]byte(tagName)) != 0
-}
+// var voidElems = map[string]struct{}{
+// 	"area":   {},
+// 	"base":   {},
+// 	"br":     {},
+// 	"col":    {},
+// 	"embed":  {},
+// 	"hr":     {},
+// 	"img":    {},
+// 	"input":  {},
+// 	"keygen": {},
+// 	"link":   {},
+// 	"meta":   {},
+// 	"param":  {},
+// 	"source": {},
+// 	"track":  {},
+// 	"wbr":    {},
+// }
 
-func isCompoTagName(tagName string, decodingSVG bool) bool {
-	if decodingSVG {
-		return false
-	}
-	return !isHTMLTagName(tagName)
-}
-
-var voidElems = map[string]struct{}{
-	"area":   {},
-	"base":   {},
-	"br":     {},
-	"col":    {},
-	"embed":  {},
-	"hr":     {},
-	"img":    {},
-	"input":  {},
-	"keygen": {},
-	"link":   {},
-	"meta":   {},
-	"param":  {},
-	"source": {},
-	"track":  {},
-	"wbr":    {},
-}
-
-func isVoidElem(tagName string) bool {
-	_, ok := voidElems[tagName]
-	return ok
-}
+// func isVoidElem(tagName string) bool {
+// 	_, ok := voidElems[tagName]
+// 	return ok
+// }
